벡터 이미지 검색 다건 조회에서 단건 조회로 변경

+++ src/main/java/com/ryu/bigdata/vo/SkuImgVo.java
@@ -7,7 +7,7 @@ public class SkuImgVo {
 
     private String vector;
     private int skuId;
-    private int skuImgId;
+    private double skuImgId;
     private String s3Path;
     private String regDateTime;
     private int numofvector;
